feat: 코스리뷰 전체보기 서비스 구현

+++ src/main/java/umc/catchy/domain/course/domain/Course.java
@@ -49,7 +49,7 @@ public class Course extends BaseTimeEntity {
     private boolean hasReview;
 
     @Setter
-    private Double rating;
+    private Float rating;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
